Rename quill to richtext, make text entry in evidence page clearer
