+++ survey/templates/survey/improvement_plan.html
@@ -57,7 +57,7 @@
                 ></div>
 
                 <h5>Evidence statement</h5>
-                <div class="sort-quill-editor mb-2"
+                <div class="sort-richtext-field mb-2"
                      data-update-url="{% url 'survey_evidence_gathering_update' survey.id evidence_section.section_id %}"
                      data-init-contents="{{ evidence_section.text }}" data-view-only="true"></div>
                 <h5>Evidence files</h5>
