+++ survey/templates/survey/report.html
@@ -25,7 +25,7 @@
 
                         {% if section.evidence %}
                             <h3>Evidence statement</h3>
-                            <div class="sort-quill-editor"
+                            <div class="sort-richtext-field"
                                  data-update-url="{% url 'survey_evidence_gathering_update' survey.id forloop.counter0 %}"
                                  data-init-contents="{{ section.evidence.text }}" data-view-only="true"></div>
                             {#            <h5>Evidence files</h5>#}
